vms/platformvm: move execconfig package content into config package + fix lint unused var
