+++ vms/platformvm/blocks/builder/helpers_test.go
@@ -40,7 +40,6 @@ import (
 	"github.com/ava-labs/avalanchego/vms/components/avax"
 	"github.com/ava-labs/avalanchego/vms/platformvm/api"
 	"github.com/ava-labs/avalanchego/vms/platformvm/config"
-	"github.com/ava-labs/avalanchego/vms/platformvm/execconfig"
 	"github.com/ava-labs/avalanchego/vms/platformvm/fx"
 	"github.com/ava-labs/avalanchego/vms/platformvm/metrics"
 	"github.com/ava-labs/avalanchego/vms/platformvm/reward"
@@ -230,7 +229,7 @@ func defaultState(
 ) state.State {
 	require := require.New(t)
 
-	execCfg, _ := execconfig.GetExecutionConfig([]byte(`{}`))
+	execCfg, _ := config.GetExecutionConfig([]byte(`{}`))
 	genesisBytes := buildGenesisTest(t, ctx)
 	state, err := state.New(
 		db,
