+++ vms/platformvm/vm_regression_test.go
@@ -32,7 +32,6 @@ import (
 	"github.com/ava-labs/avalanchego/vms/components/avax"
 	"github.com/ava-labs/avalanchego/vms/platformvm/blocks"
 	"github.com/ava-labs/avalanchego/vms/platformvm/config"
-	"github.com/ava-labs/avalanchego/vms/platformvm/execconfig"
 	"github.com/ava-labs/avalanchego/vms/platformvm/metrics"
 	"github.com/ava-labs/avalanchego/vms/platformvm/reward"
 	"github.com/ava-labs/avalanchego/vms/platformvm/signer"
@@ -656,7 +655,7 @@ func TestRejectedStateRegressionInvalidValidatorTimestamp(t *testing.T) {
 	// Force a reload of the state from the database.
 	vm.Config.Validators = validators.NewManager()
 	vm.Config.Validators.Add(constants.PrimaryNetworkID, validators.NewSet())
-	execCfg, _ := execconfig.GetExecutionConfig(nil)
+	execCfg, _ := config.GetExecutionConfig(nil)
 	is, err := state.New(
 		vm.dbManager.Current().Database,
 		nil,
@@ -968,7 +967,7 @@ func TestRejectedStateRegressionInvalidValidatorReward(t *testing.T) {
 	// Force a reload of the state from the database.
 	vm.Config.Validators = validators.NewManager()
 	vm.Config.Validators.Add(constants.PrimaryNetworkID, validators.NewSet())
-	execCfg, _ := execconfig.GetExecutionConfig(nil)
+	execCfg, _ := config.GetExecutionConfig(nil)
 	is, err := state.New(
 		vm.dbManager.Current().Database,
 		nil,
